Coded more on yes_no_v3
changed the return statement to be a list to display [True or False if invalid answer, "] My code is so complicated that I feel sorry for not following the teachers example. If you enter being a new player it asks you if you are having fun. I also moved the input question statement out of the function so that the yes_no function could be used for anything. This resulted in more code in the main routine but a function that can be used for anything and not just once.