Added round_v1
Added round_v1 which is the similar to token_generator_v2 but with some big changes. Your round, token,, and balance is displayed at the end of each  loop. The farewell message now tells you your starting and ending balance as well as thanking you for playing. The continue playing prompt was changed to typing C for continue or E for exit to <enter> to continue and X to exit